Automation/requirements.txt
paho-mqtt==1.6.1
orjson==3.10.7
